+++ tests/Fixtures/DiverseLevelThree/Modules/Customer/Models/Customer.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Fixtures\DiverseLevelThree\Modules\Customer\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+final class Customer extends Model
+{
+    protected $table = 'customers';
+}